Modified getMatStr
The function is used to generate string of the blas classes. The default class for INT64 was double. Changed that to long.

+++ src/codegen.cpp
@@ -1091,9 +1091,9 @@ std::string VCompiler::getMatTypeStr(VTypePtr vtype){
     switch(scalarType->getScalarTag()){
     
       case ScalarType::SCALAR_INT32 :
-	    matTypeStr="Double";
+	    matTypeStr="Int";
       case ScalarType::SCALAR_INT64:
-	    matTypeStr="Double";
+	    matTypeStr="Long";
       case ScalarType::SCALAR_FLOAT32:
 	    matTypeStr="Single";
       case ScalarType::SCALAR_FLOAT64:
